Improvements for working with WildFly 8 / JBoss EAP 6.x (AS 7.x)
* The JBoss Remoting JMX Connector fails to do a few things properly.
** No ConnectionNotification Support.
** Throws Unchecked RuntimeException for connect() timeouts. https://issues.jboss.org/browse/REMJMX-90
** Copies the 'handback' object to the remote server as a serialized object. https://issues.jboss.org/browse/REMJMX-91

I've filed bugs for all of these (the first was a duplicate, and reportedly in progress) and added 'workarounds' into fast-jmx for the remaining two.

If you're using any service URL containing "remoting-jmx://", the SynchronousConnectorAdapter will be used. This implements basic client-side OPEN / CLOSED notification (FAILED is not handled) and safely wraps the 'connect' to wrap any unchecked Exceptions into IOException.
To address the handback object issue, I've started identifying Connection objects by a pseudorandom type 4 UUID, and moved the collection to a HashMap. This seems to address the functional issue of JBoss trying to send serialized instances of org.collectd.Connection to the remote server as a handback object, which caused ClassNotFoundExceptions.

The result is that connection backoff seems to work well, and we can collect data from EAP 6.1 if all the proper jars are in the classpath.